requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
dataclasses-json==0.5.6
idna==3.2
marshmallow==3.13.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
pyasn1==0.4.8
PyBluez==0.23
requests==2.26.0
RPi.GPIO==0.7.0
rsa==4.7.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.6
uuid==1.30
cbor2==5.4.2.post1
typing~=3.7.4.3
pn532pi~=1.3